deleted gamestop_de,gamestop_it spiders as countries added to gamestop_ca as  renamed to gamestop
